refactor: 도메인 기반 패키지 구조로 재정리
